requirements.txt
PyYAML>=6.0.1
jsonschema>=4.22.0
pytest>=8.0.0
rank-bm25>=0.2.2
